add test data with 1 frame
