+++ source/tests/dentest/dataset/Al13Cu3/type.raw
@@ -1,0 +1,16 @@
+0
+0
+0
+0
+0
+0
+0
+0
+0
+0
+0
+0
+0
+1
+1
+1
