+++ source/tests/dentest/dataset/Al13Cu3/type_map.raw
@@ -1,0 +1,2 @@
+Al
+Cu
